SectionManager: close_objects -> ClosableObject
This fits there way better contextually. Here it gets its own little
cute test so the SectionManagerTest can radically cleaned up later.
