+++ coalib/tests/output/ClosableObjectTest.py
@@ -2,10 +2,10 @@ import sys
 import unittest
 
 sys.path.insert(0, ".")
-from coalib.output.ClosableObject import ClosableObject
+from coalib.output.ClosableObject import ClosableObject, close_objects
 
 
-class ClosablePrinterTest(unittest.TestCase):
+class ClosableObjectTest(unittest.TestCase):
     def setUp(self):
         self.uut = ClosableObject()
 
@@ -14,6 +14,11 @@ class ClosablePrinterTest(unittest.TestCase):
         self.uut.close()
         self.assertTrue(self.uut._closed)
         self.uut.close()
+        self.assertTrue(self.uut._closed)
+
+    def test_close_objects(self):
+        close_objects(None, 5)
+        close_objects(self.uut)
         self.assertTrue(self.uut._closed)
